Rollback of the "C *= 1.02" and "tagged PDF" changes by Mark Storer.
The side effects of the "C *= 1.02" change caused one bug that in the meanwhile was already fixed; but also one I couldn't fix (#2781267). I don't want to risk more bugs for only 2% gain in speed. The idea of PDF objects keeping references and vice versa is good, but should be tested more thoroughly; especially in the case of the manipulation of partially read PDFs.

As a result of the rollback, the Tagged PDF changes were also removed, because of 2 reasons:
1. the changes didn't work anymore after the rollback.
2. the changes broke existing examples: see bug #2774406

I'll have a closer look at the "Tagged PDF" classes soon, but as long as it breaks existing examples (and as long as I don't know how to fix existing examples), it can't be in the release.

git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@3912 3a709846-db2a-4fbd-be0b-eb932eee6479