Checking for the CRL URL in the signing certificate isn't sufficient.
In most of the cases, the URL is in the certificate of the issuer. Actually, I think one should add the CRLs of all the certificates 'higher' than the signing certificate.

Extra Logger in MakeSignature
CrlClient implementation to get a CRL from a specific URL.
Setting properties for CrlClientUrl
I replaced CrlClientImp and CrlClientUrl with one single class CrlClientOnline

I also normalize the hashing algorithm in case the provider is not null.
For instance: I tried with SunPKCS11 as provider and it didn't work for SHA256, only for SHA-256.
As BC accepts both, this change shouldn't cause any problems.

git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@5296 3a709846-db2a-4fbd-be0b-eb932eee6479