Made a VerificationError class and used it to create error messages when verifying certificates.
Added member variables to LtvValidation to allow better fine-tuning (not implemented yet)
Changed the method that verifies signatures in LtvValidation to make it more generic.
Implemented options to verify against a key store and to verify the complete chain.
Added the possibility to create your own CertificateVerifier in case some verification method isn't supported out of the box.
Renamed LtvValidation to LtvVerifier.
Added a way to prevent online checking of certificates (if it's part of your best practices to have all verification in the DSS).
Moved the keyStore to the super class.
Allow CRLs to be signed by a trusted anchor.
Allow OCSP responsed to be signed by a trusted anchor.
Instead of returning a boolean, the CertificateVerifiers now return a list of VerificationOK objects.
If the certificate couldn't be verified, the list will be empty.
Changed VerificationError into VerificationException
Changed LtvVerifier so that you have access to the VerificationOK info even if an exception was thrown.
Extra method that doesn't do any CRL checking
Added a VerificationOK for root certificates

git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@5452 3a709846-db2a-4fbd-be0b-eb932eee6479