This is a hack that is introduced due to a problem caused by Adobe Acrobat / Reader.
When you have more than one Author and you enter these authors like this "Author 1; Author 2" in the Info dictionary as well as in the XMP metadata, then Adobe's viewer will show quotes when looking at the metadata.
These quotes disappear if (1) you put "Author 1" and "Author 2" in different list items in the XMP metadata, and (2) add "Author 1" only in the Info dictionary.
However: if you do this in a PDF/A file, you break PDF/A conformance. Hence, this solution only works if you don't need PDF/A, but require XMP metadata.
There is no solution to remove the quotes if you need PDF/A.

See DEV-1244 and http://stackoverflow.com/a/28253095/1622493 for more info.

git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@6737 3a709846-db2a-4fbd-be0b-eb932eee6479