Make PdfReader#setPageContent method more consistent.
setPageContent used to remove recursively the old contents and reuse old xRefs, so if a content stream or an array of streams was reused across the pages or in other ways, the result of stamping was invalid.
Now by default setPageContent does not remove old content recursively. But you can use a new overload to do so.
When old content is not removed, it may be left in the resulting document. Use PdfReader#removeUnusedObjects() to clean up.
See also PdfStamperTest#setPageContentTest01, PdfStamperTest#setPageContentTest02.

DEV-1391