Reuse of PdfDestination object causes problems with links
Once you set a page on a PdfDestination, you can't change it afterwards.
This means that we need to be careful when creating different links using the same PdfDestination.
A copy constructor was introduced and it is being used to create a new PdfDestination when creating links or actions.

See https://jira.itextsupport.com/browse/DEV-1474