update font selector algorithm
Before this update 'Arial Regular' and 'Arial Light' was the same fonts: 'Arial' with Normal style. And only alphabetical order controls priority.
From now when we add regular font, we add them twice. For example 'Arial Regular'. We add them as 'Arial Regular' and as 'Arial'. So, if we try to find 'Arial' we will find 'Arial Regular'.

DEV-1508