Fixed encoding of SignedAttributeV2 field
The main problem was related to the fact that iText generated the SigningCertificateV2 attribute using BER. It happened because iText wrote AlgorithmIdentifier attribute (RFC 5035, 5.4.1.1, definition of ESSCertIDv2) even if the value was sha256 (which is default; see previous reference). If we look into X.690-0207, clause 11.5, we can see that using DER all the components of a sequence having default values shall not be included. The next problem was that iText always wrote NULL as the value for the parameters attribute (see X.590, clause 7.2, AlgorithmIdentifier attribute). As this attribute is marked as OPTIONAL it should simply be omitted

Resolves: DEV-1507
See also: SUP-948