Set version in append mode.
iText didn't set the version in append mode. By definition this shouldn't be necessary, but PdfSignatureAppearance uses PdfVersion.getVersion() to create its dictionaries. In the case of PDF/A2-a it would incorrectly use 1.4 as a version and not the provided (or original version) and as such iText generated PDFs that weren't compliant to PDF/A2-a.

SUP-1117