Fix issue with CIDSet generation
CIDSet shall be based on font.maxGlyphId property of the font, it is maxp.numGlyphs for ttf, because technically we convert all unused glyphs to space, e.g. just remove outlines.
So, we should mark as used all glyphs of the embedded font.
And now we generate CIDSet for all PDF/A documents, include PDF/A-2 and PDF/A-3.

Resolves: DEV-1615